% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIG.R
\name{GIGpack}
\alias{GIGpack}
\alias{rgig}
\alias{dgig}
\title{Generalised Inverse Gaussian}
\arguments{
\item{p}{: lambda parameter}

\item{a}{: chi parameter}

\item{b}{: psi parameter}

\item{n}{: number of observations}

\item{log_density}{: returns log density}

\item{x}{: vector of points}
}
\value{
\itemize{
\item {\code{rgig}} {returns a vector of random numbers from GIG(a,b,p).}
\item {\code{dgig}} {returns returns density of AGIG(a,b,p)at point x.}
}
}
\description{
Probability distribution function, Quantile Random generation
for the Generalised Inverse Gaussian with three-parameters a,b,p.
}
\details{
The Generalised Inverse Gaussian distrubtion(GIG), which has the following pdf:

\deqn{f(x) = x^{\lambda -1}exp{-\frac{\omega}{2}(x + \frac{1}{x})}}{f(x) = x^(\lambda -1) exp{-\omega/2 (x + 1/x)}}
}
\examples{
rgig <- function(double P, double a, double b, int n)
dgig <-  function(std::vector<double> x, double a, double b, double p,bool log_density)

}
\references{
Devroye, L. Random variate generation for the generalized inverse Gaussian distribution.
Stat Comput 24, 239–246 (2014). https://doi.org/10.1007/s11222-012-9367-z
}
